% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{latlng_to_mgrs}
\alias{latlng_to_mgrs}
\title{Convert latitude/longitude to MGRS string}
\usage{
latlng_to_mgrs(latitude, longitude, degrees = TRUE, precision = 5L)
}
\arguments{
\item{latitude, longitude}{coordinates}

\item{degrees}{are latitude/longitude in degrees? Default: \code{TRUE}}

\item{precision}{0:5; level of precision for the conversion. Default \code{5}}
}
\description{
Convert latitude/longitude to MGRS string
}
\examples{
latlng_to_mgrs(42, -93)
}
