% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgrs_to_latlng}
\alias{mgrs_to_latlng}
\title{Convert an MGRS string to latitude/longitude}
\usage{
mgrs_to_latlng(MGRS, degrees = TRUE)
}
\arguments{
\item{MGRS}{an MGRS string}

\item{degrees}{convert to degrees? Default: \code{TRUE}}
}
\value{
\code{data.frame}
}
\description{
Convert an MGRS string to latitude/longitude
}
\note{
vectorized
}
\examples{
mgrs_to_latlng("15TWG0000049776")
}
