% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgrs_to_ups}
\alias{mgrs_to_ups}
\title{Convert MGRS to UPS}
\usage{
mgrs_to_ups(mgrs_string, include_mgrs_ref = TRUE)
}
\arguments{
\item{mgrs_string}{a character vector of MGRS strings}

\item{include_mgrs_ref}{if \code{TRUE} the data frame returned fill include
the MGRS reference in a column named \code{mgrs}. Default: \code{TRUE}.}
}
\value{
\code{data.frame}
}
\description{
Convert MGRS to UPS
}
\note{
vectorized
}
\examples{
mgrs_to_ups("ZGC2677330125")
}
