% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgrs_to_utm}
\alias{mgrs_to_utm}
\title{Convert MGRS to UTM}
\usage{
mgrs_to_utm(mgrs_string)
}
\arguments{
\item{MGRS}{an MGRS string}
}
\value{
\code{data.frame}
}
\description{
Convert MGRS to UTM
}
\examples{
mgrs_to_utm("48PUV7729883034")
}
