% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ups_to_latlng}
\alias{ups_to_latlng}
\title{Convert UPS to Latitude/Longitude}
\usage{
ups_to_latlng(hemisphere, easting, northing, degrees = TRUE)
}
\arguments{
\item{hemisphere}{South (\code{S}) or North (\code{N})}

\item{easting, northing}{easting (X) / northing (Y) (meters)}

\item{degrees}{convert to degrees? Default: \code{TRUE}}
}
\description{
Convert UPS to Latitude/Longitude
}
\examples{
ups_to_latlng("N", 2426773, 1530125)
}
