% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ups_to_mgrs}
\alias{ups_to_mgrs}
\title{Convert UPS to MGRS}
\usage{
ups_to_mgrs(hemisphere, easting, northing, precision = 5L)
}
\arguments{
\item{hemisphere}{South (\code{S}) or North (\code{N})}

\item{easting, northing}{easting (X) / northing (Y) (meters)}

\item{precision}{0:5; level of precision for the conversion. Default \code{5}}
}
\description{
Convert UPS to MGRS
}
\examples{
ups_to_mgrs("N", 2426773, 1530125)
}
