% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{utm_to_latlng}
\alias{utm_to_latlng}
\title{Convert UTM to Latitude/Longitude}
\usage{
utm_to_latlng(zone, hemisphere, easting, northing, degrees = TRUE)
}
\arguments{
\item{zone}{1:60}

\item{hemisphere}{South (\code{S}) or North (\code{N})}

\item{easting, northing}{easting (X) / northing (Y) (meters)}

\item{degrees}{convert to degrees? Default: \code{TRUE}}
}
\description{
Convert UTM to Latitude/Longitude
}
\examples{
utm_to_latlng(48, "N", 377299, 1483035)
}
