% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{utm_to_mgrs}
\alias{utm_to_mgrs}
\title{Convert UTM to MGRS}
\usage{
utm_to_mgrs(zone, hemisphere, easting, northing, precision = 5L)
}
\arguments{
\item{zone}{1:60}

\item{hemisphere}{South (\code{S}) or North (\code{N})}

\item{easting, northing}{easting (X) / northing (Y) (meters)}

\item{precision}{0:5; level of precision for the conversion. Default \code{5}}
}
\description{
Convert UTM to MGRS
}
\examples{
utm_to_mgrs(48, "N", 377299, 1483035)
}
