% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgrs-package.R
\docType{package}
\name{mgrs}
\alias{mgrs}
\alias{mgrs-package}
\title{Convert 'MGRS' Coordinates To and From Other Coordiante Systems}
\description{
Convert 'MGRS' Coordinates To and From Other Coordiante Systems
}
\author{
Bob Rudis (bob@rud.is)
}
