% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgrs_to_ups}
\alias{mgrs_to_ups}
\title{Convert MGRS to UPS}
\usage{
mgrs_to_ups(mgrs_string)
}
\arguments{
\item{MGRS}{an MGRS string}
}
\value{
\code{data.frame}
}
\description{
Convert MGRS to UPS
}
\examples{
mgrs_to_ups("ZGC2677330125")
}
