% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgrs_to_utm}
\alias{mgrs_to_utm}
\title{Convert MGRS to UTM}
\usage{
mgrs_to_utm(mgrs_string, include_mgrs_ref = TRUE)
}
\arguments{
\item{mgrs_string}{acharacter vector of MGRS strings}

\item{include_mgrs_ref}{if \code{TRUE} the data frame returned fill include
the MGRS reference in a column named \code{mgrs}. Default: \code{TRUE}.}
}
\value{
\code{data.frame}
}
\description{
Convert MGRS to UTM
}
\note{
vectorized
}
\examples{
mgrs_to_utm("48PUV7729883034")
}
